refactor: fold statement warnings into planner_warnings, coordinate seq-scan warning
Move the per-statement warning machinery (thread-locals, dedup helper,
message entry points) into planner_warnings and drop the separate
statement_warnings module. Suppress the generic sequential-scan warning
when a join- or aggregate-scan planner warning already fired for the
statement, since that names a concrete alternative. Regenerate the
affected regress expected outputs.
