+++ pg_search/src/postgres/planner_warnings.rs
@@ -15,8 +15,11 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-use std::cell::RefCell;
+use std::cell::{Cell, RefCell};
 use std::collections::BTreeSet;
+use std::thread::LocalKey;
+
+use pgrx::pg_sys;
 
 use crate::api::HashMap;
 
@@ -229,4 +232,85 @@ pub fn emit_planner_warnings() {
             }
         }
     });
+}
+
+//
+// Per-statement execution-time warnings.
+//
+// Unlike the planner warnings above, which are batched and flushed at the end of planning, these
+// fire during execution and dedup against the currently-executing statement so a warning raised
+// from many `@@@` predicate nodes surfaces just once.
+//
+
+/// Uniquely identifies the currently-executing statement (its start timestamp).
+type StatementId = pg_sys::TimestampTz;
+
+/// Sentinel [`StatementId`] that never matches a real statement, for initializing trackers.
+const NEVER: StatementId = pg_sys::TimestampTz::MIN;
+
+thread_local! {
+    /// The statement for which we last warned that the table is being sequentially scanned / that
+    /// the match set spilled. A statement can have many `@@@` predicate nodes, each with its own
+    /// per-`fcinfo` cache, so dedup at the statement level to warn just once.
+    static WARNED_SEQ_SCAN_AT: Cell<StatementId> = const { Cell::new(NEVER) };
+    static WARNED_SPILLED_AT: Cell<StatementId> = const { Cell::new(NEVER) };
+}
+
+/// Identifies the currently-executing statement. Distinct statements -- including each `EXECUTE` of
+/// a prepared statement -- get distinct values, letting us dedup a warning to once per statement.
+fn current_statement_id() -> StatementId {
+    unsafe { pg_sys::GetCurrentStatementStartTimestamp() }
+}
+
+/// Emit `message` as a WARNING at most once per statement, tracked by `warned_at`.
+fn warn_once_per_statement(warned_at: &'static LocalKey<Cell<StatementId>>, message: &str) {
+    let stmt_id = current_statement_id();
+    if warned_at.with(|last| last.replace(stmt_id)) != stmt_id {
+        pgrx::warning!("{message}");
+    }
+}
+
+/// Planner-warning categories (the custom scans' `NAME`s) whose presence means the user already
+/// received more actionable guidance than the generic sequential-scan warning.
+const JOIN_SCAN_CATEGORY: &str = "ParadeDB Join Scan";
+const AGGREGATE_SCAN_CATEGORY: &str = "ParadeDB Aggregate Scan";
+
+/// Whether a warning was recorded for the current planning cycle under any of `categories`. The
+/// planner state is populated during planning and only cleared at the start of the next planning
+/// cycle, so it is still readable while the plan executes.
+fn planner_warned_in(categories: &[&str]) -> bool {
+    PLANNER_STATE.with(|state| {
+        let state = state.borrow();
+        categories.iter().any(|cat| {
+            state
+                .warnings
+                .get(*cat)
+                .is_some_and(|msgs| !msgs.is_empty())
+        })
+    })
+}
+
+/// Warn (once per statement) that a `@@@` predicate is being applied as a per-row filter over a
+/// sequential scan rather than via the index.
+///
+/// Suppressed when the planner already emitted a join- or aggregate-scan warning for this
+/// statement: that warning names a concrete alternative, so the generic sequential-scan warning
+/// would just be noise next to it.
+pub fn warn_sequential_scan() {
+    if planner_warned_in(&[JOIN_SCAN_CATEGORY, AGGREGATE_SCAN_CATEGORY]) {
+        return;
+    }
+    warn_once_per_statement(
+        &WARNED_SEQ_SCAN_AT,
+        "the table is being sequentially scanned for this query, so performance may be slow",
+    );
+}
+
+/// Warn (once per statement) that the materialized filter match set exceeded `work_mem` and
+/// spilled to a temporary file.
+pub fn warn_filter_spilled() {
+    warn_once_per_statement(
+        &WARNED_SPILLED_AT,
+        "the query's filter match set exceeded work_mem and spilled to a temporary file; query performance may be degraded",
+    );
 }
