+++ pg_search/tests/pg_regress/expected/issue_2533.out
@@ -163,7 +163,6 @@ SELECT (SELECT COUNT(*)
            OR ((users.color @@@ 'blue') AND (users.name @@@ 'bob')) AND (products.color @@@ 'blue')
            OR (NOT (products.name @@@ 'bob')));
 WARNING:  JoinScan not used: aggregates are not supported. Enable paradedb.enable_aggregate_custom_scan to optimize this query. (tables: products, users)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  count | count 
 -------+-------
     13 |    13
@@ -188,7 +187,6 @@ SELECT (SELECT COUNT(*)
            OR ((users.color @@@ 'blue') OR (NOT (users.name @@@ 'bob'))) AND
               NOT ((orders.name @@@ 'bob') OR (orders.age @@@ '20')) AND (orders.name @@@ 'bob'));
 WARNING:  JoinScan not used: aggregates are not supported. Enable paradedb.enable_aggregate_custom_scan to optimize this query. (tables: orders, users)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  count | count 
 -------+-------
      0 |     0
@@ -213,7 +211,6 @@ SELECT (SELECT COUNT(*)
            OR ((users.name @@@ 'bob') OR (NOT (users.color @@@ 'blue'))) AND NOT (products.color @@@ 'blue') AND
               ((products.color @@@ 'blue') OR (products.color @@@ 'blue')));
 WARNING:  JoinScan not used: aggregates are not supported. Enable paradedb.enable_aggregate_custom_scan to optimize this query. (tables: products, users)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  count | count 
 -------+-------
      2 |     2
@@ -238,7 +235,6 @@ SELECT (SELECT COUNT(*)
            OR ((users.age @@@ '20') OR (NOT (users.color @@@ 'blue'))) AND NOT (products.age @@@ '20') AND
               (NOT (NOT (products.name @@@ 'bob'))));
 WARNING:  JoinScan not used: aggregates are not supported. Enable paradedb.enable_aggregate_custom_scan to optimize this query. (tables: products, users)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  count | count 
 -------+-------
      3 |     3
@@ -263,7 +259,6 @@ SELECT (SELECT COUNT(*)
            OR ((NOT (users.color @@@ 'blue')) OR (users.color @@@ 'blue')) AND NOT (products.age @@@ '20') AND
               ((products.color @@@ 'blue') AND (products.name @@@ 'bob')));
 WARNING:  JoinScan not used: aggregates are not supported. Enable paradedb.enable_aggregate_custom_scan to optimize this query. (tables: products, users)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  count | count 
 -------+-------
      2 |     2
@@ -286,7 +281,6 @@ SELECT (SELECT COUNT(*)
               NOT NOT (NOT ((NOT (orders.name @@@ 'bob')) OR (orders.name @@@ 'bob'))) AND NOT (users.age @@@ '20')
            OR ((users.age @@@ '20') AND (NOT (users.color @@@ 'blue'))));
 WARNING:  JoinScan not used: aggregates are not supported. Enable paradedb.enable_aggregate_custom_scan to optimize this query. (tables: orders, users)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  count | count 
 -------+-------
     13 |    13
@@ -309,7 +303,6 @@ SELECT (SELECT COUNT(*)
               NOT NOT ((NOT (orders.color @@@ 'blue')) OR (NOT (orders.color @@@ 'blue'))) AND NOT (users.age @@@ '20')
            OR ((NOT (users.color @@@ 'blue')) AND (users.color @@@ 'blue')));
 WARNING:  JoinScan not used: only INNER, ANTI, and SEMI JOIN are currently supported (types: LEFT, RIGHT) (tables: orders, users)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  count | count 
 -------+-------
     13 |    13
@@ -334,7 +327,6 @@ SELECT (SELECT COUNT(*)
            OR (orders.color @@@ 'blue') AND NOT (users.age @@@ '20')
            OR ((users.name @@@ 'bob') AND (NOT (users.color @@@ 'blue'))));
 WARNING:  JoinScan not used: aggregates are not supported. Enable paradedb.enable_aggregate_custom_scan to optimize this query. (tables: orders, users)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  count | count 
 -------+-------
     14 |    14
@@ -356,7 +348,6 @@ SELECT (SELECT COUNT(*)
         WHERE (NOT (products.id @@@ '3'))
            OR ((products.name @@@ 'bob') AND (orders.id @@@ '3')));
 WARNING:  JoinScan not used: aggregates are not supported. Enable paradedb.enable_aggregate_custom_scan to optimize this query. (tables: orders, products)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  count | count 
 -------+-------
     19 |    19
