+++ pg_search/tests/pg_regress/expected/issue_3827.out
@@ -78,7 +78,6 @@ GROUP BY txt
 HAVING (txt @@@ pdb.parse('foo')) OR SUM(n) < 0
 ORDER BY txt;
 WARNING:  Aggregate Scan not used: grouping column txt exists, but is not a fast field. To disable this warning: SET paradedb.check_aggregate_scan = false (table: issue_3827_t)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  txt 
 -----
  foo
