+++ pg_search/tests/pg_regress/expected/join_predicates.out
@@ -170,7 +170,6 @@ JOIN suppliers s ON p.supplier_id = s.id
 WHERE p.description @@@ 'wireless' OR s.contact_info @@@ 'wireless'
 ORDER BY p.id;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: p, s)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  id  |      name      | supplier_name 
 -----+----------------+---------------
  201 | Wireless Mouse | TechCorp
