+++ pg_search/tests/pg_regress/expected/join_tests.out
@@ -131,7 +131,6 @@ INNER JOIN books b ON a.id = b.author_id
 WHERE (a.bio @@@ 'science' OR b.content @@@ 'technology')
 ORDER BY a.id, b.id, author_score DESC, book_score DESC;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
   author_name  |      book_title      | author_score | book_score 
 ---------------+----------------------+--------------+------------
  J.K. Rowling  | Harry Potter Magic   |            0 |    0.98641
@@ -152,7 +151,6 @@ LEFT JOIN books b ON a.id = b.author_id
 WHERE (a.bio @@@ 'mystery' OR b.content @@@ 'romance')
 ORDER BY a.id, b.id, author_score DESC, book_score DESC;
 WARNING:  JoinScan not used: only INNER, ANTI, and SEMI JOIN are currently supported (types: LEFT, RIGHT) (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
    author_name   |     book_title      | author_score | book_score 
 -----------------+---------------------+--------------+------------
  Agatha Christie | Murder Mystery Case |    1.5552412 |          0
@@ -169,7 +167,6 @@ RIGHT JOIN books b ON a.id = b.author_id
 WHERE (a.bio @@@ 'fiction' OR b.content @@@ 'magic')
 ORDER BY a.id, b.id, author_score DESC, book_score DESC;
 WARNING:  JoinScan not used: only INNER, ANTI, and SEMI JOIN are currently supported (types: LEFT, RIGHT) (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  author_name  |      book_title      | author_score | book_score 
 --------------+----------------------+--------------+------------
  J.K. Rowling | Harry Potter Magic   |            0 |  1.3025584
@@ -207,7 +204,6 @@ WHERE (a.bio @@@ 'author' OR b.content @@@ 'mystery')
 ORDER BY a.id, b.id, author_score DESC, book_score DESC
 LIMIT 10;
 WARNING:  JoinScan not used: at least one equi-join key (e.g., a.id = b.id) is required (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  author_name  |      book_title      | author_score | book_score 
 --------------+----------------------+--------------+------------
  J.K. Rowling | Harry Potter Horrors |   0.66167223 |          0
@@ -233,7 +229,6 @@ INNER JOIN books b ON a.birth_year < b.publication_year
 WHERE (a.bio @@@ 'fiction' OR b.content @@@ 'love')
 ORDER BY a.id, b.id, author_score DESC, book_score DESC;
 WARNING:  JoinScan not used: at least one equi-join key (e.g., a.id = b.id) is required (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  author_name  |      book_title      | author_score | book_score 
 --------------+----------------------+--------------+------------
  Stephen King | Harry Potter Magic   |   0.66167223 |          0
@@ -271,7 +266,6 @@ INNER JOIN books b ON a.birth_year + 50 > b.publication_year
 WHERE (a.bio @@@ 'writer' OR b.content @@@ 'programming')
 ORDER BY a.id, b.id, author_score DESC, book_score DESC;
 WARNING:  JoinScan not used: at least one equi-join key (e.g., a.id = b.id) is required (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
    author_name   |     book_title      | author_score | book_score 
 -----------------+---------------------+--------------+------------
  Agatha Christie | Murder Mystery Case |    1.0395092 |          0
@@ -308,7 +302,6 @@ WHERE (a.bio @@@ 'smartphone' OR b.content @@@ 'performance')
 ORDER BY a.id, b.id, author_score DESC, book_score DESC
 LIMIT 10;
 WARNING:  JoinScan not used: at least one equi-join key (e.g., a.id = b.id) is required (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
    author_name   |                        book_content                         | author_score | book_score 
 -----------------+-------------------------------------------------------------+--------------+------------
  J.K. Rowling    | A horror story about supernatural terror events performance |            0 |  1.3702781
@@ -365,7 +358,6 @@ WHERE (a.bio @@@ 'smartphone' OR a.country @@@ 'British' OR b.content @@@ 'perfo
 ORDER BY a.id, b.id, author_score DESC, book_score DESC
 LIMIT 10;
 WARNING:  JoinScan not used: at least one equi-join key (e.g., a.id = b.id) is required (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
    author_name   | author_country |                        book_content                         | author_score | book_score 
 -----------------+----------------+-------------------------------------------------------------+--------------+------------
  J.K. Rowling    | UK             | A horror story about supernatural terror events performance |            0 |  1.3702781
@@ -394,7 +386,6 @@ INNER JOIN books b ON a.id = b.author_id
 WHERE (a.bio @@@ 'science' OR b.content @@@ 'mystery' OR b.price > 25.00)
 ORDER BY a.id, b.id, author_score DESC, book_score DESC;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
    author_name   |      book_title      | author_score | book_score 
 -----------------+----------------------+--------------+------------
  J.K. Rowling    | Harry Potter Magic   |            0 |          0
@@ -418,7 +409,6 @@ WHERE (a.bio @@@ 'smartphone' AND a.birth_year > 1950)
    OR (b.content @@@ 'magic' AND b.publication_year > 1980)
 ORDER BY a.id, b.id, author_score DESC, book_score DESC;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  author_name  |      book_title      | author_score | book_score 
 --------------+----------------------+--------------+------------
  J.K. Rowling | Harry Potter Magic   |            0 |  1.3025584
@@ -499,7 +489,6 @@ LEFT JOIN books b ON a.id = b.author_id
 WHERE a.bio @@@ 'author' OR b.content @@@ 'story'
 ORDER BY a.id, b.id;
 WARNING:  JoinScan not used: only INNER, ANTI, and SEMI JOIN are currently supported (types: LEFT, RIGHT) (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
    author_name   |      book_title      | author_score | book_score 
 -----------------+----------------------+--------------+------------
  J.K. Rowling    | Harry Potter Magic   |   0.66167223 | 0.27030534
@@ -611,7 +600,6 @@ WHERE (a.bio @@@ 'smartphone' OR a.birth_year = b.publication_year)
 ORDER BY a.id, b.id, author_score DESC, book_score DESC
 LIMIT 5;
 WARNING:  JoinScan not used: at least one equi-join key (e.g., a.id = b.id) is required (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  author_name |      book_title      | author_score | book_score 
 -------------+----------------------+--------------+------------
  John Smith  | Harry Potter Magic   |    1.6503524 |          0
@@ -658,7 +646,6 @@ WHERE (a.bio @@@ 'smartphone')    -- Safe condition on authors
    OR (c.description @@@ 'electronic')    -- External condition on categories
 ORDER BY a.id, c.id, author_score DESC, category_score DESC;
 WARNING:  JoinScan not used: at least one equi-join key (e.g., a.id = b.id) is required (tables: a, c)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  author_name | category_name | author_score | category_score 
 -------------+---------------+--------------+----------------
  John Smith  | Technology    |    1.6503524 |              0
