+++ pg_search/tests/pg_regress/expected/not-paradedb_all.out
@@ -68,7 +68,6 @@ WHERE a.id = b.id AND NOT a.id @@@ paradedb.all()
    OR b.id @@@ paradedb.all()
 ORDER BY a.id, b.id;
 WARNING:  JoinScan not used: at least one equi-join key (e.g., a.id = b.id) is required (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  id | id 
 ----+----
   1 |  1
