+++ pg_search/tests/pg_regress/expected/score_join_predicates.out
@@ -69,7 +69,6 @@ JOIN authors a ON b.author_id = a.id
 WHERE (b.content @@@ 'test' OR a.name @@@ 'Rowling') AND a.age @@@ '>50'
 ORDER BY b.id, a.id;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  book_id |   author_name   | author_score | book_score 
 ---------+-----------------+--------------+------------
        1 | J.K. Rowling    |     2.540445 |          0
@@ -150,7 +149,6 @@ JOIN authors a ON b.author_id = a.id
 WHERE (a.name @@@ 'King' OR b.content @@@ 'scoring')
 ORDER BY b.id, a.id;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  book_id | author_name  | author_score | book_score 
 ---------+--------------+--------------+------------
        1 | Stephen King |    1.5404451 |          0
@@ -178,7 +176,6 @@ JOIN authors a ON b.author_id = a.id
 WHERE (a.name @@@ 'King' OR b.content @@@ 'scoring') OR a.age > 60
 ORDER BY b.id, a.id;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  book_id |   author_name   | author_score | book_score 
 ---------+-----------------+--------------+------------
        1 | Stephen King    |     2.540445 |          0
@@ -232,7 +229,6 @@ LEFT JOIN authors a ON b.author_id = a.id
 WHERE (b.content @@@ 'test' OR a.name @@@ 'Rowling') AND a.age @@@ '>50'
 ORDER BY b.id, a.id;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  book_id |   author_name   | author_score | book_score 
 ---------+-----------------+--------------+------------
        1 | J.K. Rowling    |     2.540445 |          0
@@ -269,7 +265,6 @@ JOIN authors a ON b.author_id = a.id
 WHERE (b.content @@@ 'function' OR a.name @@@ 'King') AND a.age @@@ '>50'
 ORDER BY b.id, a.id;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  book_id | author_name  | author_score_1 | author_score_2 | book_score_1 | book_score_2 
 ---------+--------------+----------------+----------------+--------------+--------------
        1 | J.K. Rowling |              1 |              1 |   0.35745716 |   0.35745716
@@ -306,7 +301,6 @@ JOIN authors a ON b.author_id = a.id
 WHERE (b.content @@@ 'test' OR a.name @@@ 'Rowling') AND a.age @@@ '>50'
 ORDER BY b.id, a.id;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  book_id |   author_name   | author_score |   author_snippet    | book_score |                                                                       book_snippet                                                                       
 ---------+-----------------+--------------+---------------------+------------+----------------------------------------------------------------------------------------------------------------------------------------------------------
        1 | J.K. Rowling    |     2.540445 | J.K. <b>Rowling</b> |          0 | 
@@ -324,7 +318,6 @@ LEFT JOIN authors a ON b.author_id = a.id
 WHERE (a.name @@@ 'King' OR b.content @@@ 'scoring')
 ORDER BY b.id, a.id;
 WARNING:  JoinScan not used: only INNER, ANTI, and SEMI JOIN are currently supported (types: LEFT, RIGHT) (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  id |     name     | author_score | book_score 
 ----+--------------+--------------+------------
   1 | Stephen King |    1.5404451 |          0
@@ -338,7 +331,6 @@ RIGHT JOIN authors a ON b.author_id = a.id
 WHERE (a.name @@@ 'King' OR b.content @@@ 'scoring')
 ORDER BY a.id;
 WARNING:  JoinScan not used: only INNER, ANTI, and SEMI JOIN are currently supported (types: LEFT, RIGHT) (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  id |     name     | author_score | book_score 
 ----+--------------+--------------+------------
   1 | Stephen King |    1.5404451 |          0
