+++ pg_search/tests/pg_regress/expected/sequential_scan.out
@@ -32,7 +32,6 @@ WARNING:  Aggregate Scan not used: query does not imply the partial index predic
 
 SELECT count(*) FROM sequential_scan WHERE body ||| 'keyword';
 WARNING:  Aggregate Scan not used: query does not imply the partial index predicate. To disable this warning: SET paradedb.check_aggregate_scan = false (table: sequential_scan)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
 WARNING:  the query's filter match set exceeded work_mem and spilled to a temporary file; query performance may be degraded
  count 
 -------
@@ -74,7 +73,6 @@ WARNING:  Aggregate Scan not used: query does not imply the partial index predic
 
 SELECT count(*) FROM sequential_scan WHERE NOT (body ||| 'keyword');
 WARNING:  Aggregate Scan not used: query does not imply the partial index predicate. To disable this warning: SET paradedb.check_aggregate_scan = false (table: sequential_scan)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
 WARNING:  the query's filter match set exceeded work_mem and spilled to a temporary file; query performance may be degraded
  count 
 -------
@@ -93,7 +91,6 @@ WARNING:  Aggregate Scan not used: query does not imply the partial index predic
 
 SELECT count(*) FROM sequential_scan WHERE body ||| 'nonexistentterm';
 WARNING:  Aggregate Scan not used: query does not imply the partial index predicate. To disable this warning: SET paradedb.check_aggregate_scan = false (table: sequential_scan)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  count 
 -------
      0
@@ -112,7 +109,6 @@ WARNING:  Aggregate Scan not used: query does not imply the partial index predic
 
 SELECT count(*) FROM sequential_scan WHERE body ||| 'keyword';
 WARNING:  Aggregate Scan not used: query does not imply the partial index predicate. To disable this warning: SET paradedb.check_aggregate_scan = false (table: sequential_scan)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  count 
 -------
  20000
