+++ pg_search/tests/pg_regress/expected/snippet_position_01_advanced.out
@@ -70,7 +70,6 @@ JOIN authors a ON b.author_id = a.id
 WHERE b.content @@@ 'test' OR a.name @@@ 'Rowling'
 ORDER BY b.id, a.id;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  book_id | author_id |   author_snippet    | author_positions | author_score | book_score 
 ---------+-----------+---------------------+------------------+--------------+------------
        1 |         1 | J.K. <b>Rowling</b> | {{5,12}}         |    1.5404451 |          0
@@ -93,7 +92,6 @@ JOIN authors a ON b.author_id = a.id
 WHERE b.content @@@ 'test' OR NOT(a.name @@@ 'Rowling')
 ORDER BY b.id, a.id;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  book_id | author_id |   author_snippet    | author_positions | author_score | book_score 
 ---------+-----------+---------------------+------------------+--------------+------------
        1 |         1 | J.K. <b>Rowling</b> | {{5,12}}         |            0 |          0
@@ -116,7 +114,6 @@ JOIN authors a ON b.author_id = a.id
 WHERE NOT(b.content @@@ 'test') OR a.name @@@ 'Rowling'
 ORDER BY b.id, a.id;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  book_id | author_id |   author_snippet    | author_positions | author_score | book_score 
 ---------+-----------+---------------------+------------------+--------------+------------
        1 |         1 | J.K. <b>Rowling</b> | {{5,12}}         |    1.5404451 |          0
@@ -135,7 +132,6 @@ JOIN authors a ON b.author_id = a.id
 WHERE NOT(b.content @@@ 'test') OR NOT(a.name @@@ 'Rowling')
 ORDER BY b.id, a.id;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  book_id | author_id | author_snippet | author_positions | author_score | book_score 
 ---------+-----------+----------------+------------------+--------------+------------
        1 |         2 |                |                  |            0 |          0
@@ -261,7 +257,6 @@ WHERE b.content @@@ 'test'
     OR r.review @@@ 'snippet'
 ORDER BY b.id, r.id;
 WARNING:  JoinScan not used: only INNER, ANTI, and SEMI JOIN are currently supported (types: LEFT, RIGHT) (tables: a, b, r)
-WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  book_id |                                                                       book_snippet                                                                       |         book_positions          |   author_snippet    | author_positions |                                               review_snippet                                                |         review_positions         | book_score | author_score | review_score 
 ---------+----------------------------------------------------------------------------------------------------------------------------------------------------------+---------------------------------+---------------------+------------------+-------------------------------------------------------------------------------------------------------------+----------------------------------+------------+--------------+--------------
        1 |                                                                                                                                                          |                                 | J.K. <b>Rowling</b> | {{5,12}}         | This is a <b>test</b> review of the <b>snippet</b> function with multiple <b>test</b> words                 | {{10,14},{29,36},{60,64}}        |          0 |    1.5404451 |     0.494645
